금강ui/컨텐츠 자동화 파이프 라인/부동산 주제 관련/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
psycopg2-binary==2.9.9
geopy==2.4.0
pandas==2.1.2
numpy==1.26.1
python-dateutil==2.8.2
